+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+1.0.0 / 2016-01-01
+==================
+  * Escape only single quotes. Before: `template:'\'\"'`, Now: `template:'\'"'`
+  * Checked support of Angular2.0 templates like `\<a \[router-link\]="\['/search'\]"\>Search\</a\>`
+
+0.2.1 / 2015-12-31
+==================
+  * Bug fix: remove \`templateUrl\` support
+
 0.2.0 / 2015-12-31
 ==================
   * Added ES6 template string quotes \` support (templateUrl: \`/path/to/template.html\`)